Move GCETranslator out from utils.go
